Rename to bypassrestrictions
