+++ src/main/java/org/owasp/webgoat/lessons/bypassrestrictions/BypassRestrictions.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.bypass_restrictions;
+package org.owasp.webgoat.lessons.bypassrestrictions;
 
 import org.owasp.webgoat.container.lessons.Category;
 import org.owasp.webgoat.container.lessons.Lesson;
