+++ src/main/java/org/owasp/webgoat/lessons/bypassrestrictions/BypassRestrictionsFrontendValidation.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.bypass_restrictions;
+package org.owasp.webgoat.lessons.bypassrestrictions;
 
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
 import org.owasp.webgoat.container.assignments.AttackResult;
@@ -30,24 +30,40 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-public class BypassRestrictionsFieldRestrictions extends AssignmentEndpoint {
+public class BypassRestrictionsFrontendValidation extends AssignmentEndpoint {
 
-    @PostMapping("/BypassRestrictions/FieldRestrictions")
+    @PostMapping("/BypassRestrictions/frontendValidation")
     @ResponseBody
-    public AttackResult completed(@RequestParam String select, @RequestParam String radio, @RequestParam String checkbox, @RequestParam String shortInput, @RequestParam String readOnlyInput) {
-        if (select.equals("option1") || select.equals("option2")) {
+    public AttackResult completed(@RequestParam String field1, @RequestParam String field2, @RequestParam String field3, @RequestParam String field4, @RequestParam String field5, @RequestParam String field6, @RequestParam String field7, @RequestParam Integer error) {
+        final String regex1 = "^[a-z]{3}$";
+        final String regex2 = "^[0-9]{3}$";
+        final String regex3 = "^[a-zA-Z0-9 ]*$";
+        final String regex4 = "^(one|two|three|four|five|six|seven|eight|nine)$";
+        final String regex5 = "^\\d{5}$";
+        final String regex6 = "^\\d{5}(-\\d{4})?$";
+        final String regex7 = "^[2-9]\\d{2}-?\\d{3}-?\\d{4}$";
+        if (error > 0) {
             return failed(this).build();
         }
-        if (radio.equals("option1") || radio.equals("option2")) {
+        if (field1.matches(regex1)) {
             return failed(this).build();
         }
-        if (checkbox.equals("on") || checkbox.equals("off")) {
+        if (field2.matches(regex2)) {
             return failed(this).build();
         }
-        if (shortInput.length() <= 5) {
+        if (field3.matches(regex3)) {
             return failed(this).build();
         }
-        if ("change".equals(readOnlyInput)) {
+        if (field4.matches(regex4)) {
+            return failed(this).build();
+        }
+        if (field5.matches(regex5)) {
+            return failed(this).build();
+        }
+        if (field6.matches(regex6)) {
+            return failed(this).build();
+        }
+        if (field7.matches(regex7)) {
             return failed(this).build();
         }
         return success(this).build();
