+++ src/main/resources/lessons/bypassrestrictions/css/bypass-restrictions.css
