+++ src/main/resources/lessons/bypassrestrictions/documentation/BypassRestrictions_FrontendValidation.adoc
