+++ src/main/resources/lessons/bypassrestrictions/documentation/BypassRestrictions_Intro.adoc
