+++ src/main/resources/lessons/bypassrestrictions/i18n/WebGoatLabels.properties
